vedio bor yuqligiga tekshirish

+++ src/main/java/metube/com/intercationservice/service/commit/CommitServiceImpl.java
@@ -1,8 +1,10 @@
 package metube.com.intercationservice.service.commit;
 
 import lombok.RequiredArgsConstructor;
+import metube.com.intercationservice.clients.VideoServiceClient;
 import metube.com.intercationservice.domian.dto.request.CommitReq;
 import metube.com.intercationservice.domian.dto.response.CommitRes;
+import metube.com.intercationservice.domian.dto.response.VideoResponse;
 import metube.com.intercationservice.domian.entity.CommitEntity;
 import metube.com.intercationservice.exception.BaseException;
 import metube.com.intercationservice.repository.CommitRepository;
@@ -17,8 +19,15 @@ import java.util.UUID;
 @RequiredArgsConstructor
 public class CommitServiceImpl implements CommitService {
     private final CommitRepository commitRepository;
+    private final VideoServiceClient videoServiceClient;
     @Override
     public CommitRes createCommit(CommitReq commitReq) {
+
+        VideoResponse videoResponse = videoServiceClient.getVideo(commitReq.getVideoId());
+        if (videoResponse == null) {
+            throw new BaseException("Video not found", HttpStatus.NOT_FOUND.value());
+        }
+
         CommitEntity build = CommitEntity.builder()
                 .userId(commitReq.getUserId())
                 .comment(commitReq.getComment())
@@ -39,6 +48,13 @@ public class CommitServiceImpl implements CommitService {
     public CommitRes findById(UUID id) {
         CommitEntity commit = commitRepository.findById(id)
                 .orElseThrow(() -> new BaseException("Commit not found", HttpStatus.NOT_FOUND.value()));
+
+        //vedio tekshirish
+        VideoResponse videoResponse = videoServiceClient.getVideo(commit.getVideoId());
+        if (videoResponse == null) {
+            throw new BaseException("Video not found", HttpStatus.NOT_FOUND.value());
+        }
+
         return CommitRes.builder()
                 .id(commit.getId())
                 .userId(commit.getUserId())
@@ -58,6 +74,12 @@ public class CommitServiceImpl implements CommitService {
         CommitEntity commit = commitRepository.findById(id)
                 .orElseThrow(() -> new BaseException("Commit not found", HttpStatus.NOT_FOUND.value()));
 
+        //vedio tekshirish
+        VideoResponse videoResponse = videoServiceClient.getVideo(commit.getVideoId());
+        if (videoResponse == null) {
+            throw new BaseException("Video not found", HttpStatus.NOT_FOUND.value());
+        }
+
         commit.setComment(commitReq.getComment());
 
         commitRepository.save(commit);
@@ -65,7 +87,14 @@ public class CommitServiceImpl implements CommitService {
 
     @Override
     public void deleteCommit(UUID id) {
-        findById(id);
+        CommitRes byId = findById(id);
+
+        //vedio tekshirish
+        VideoResponse videoResponse = videoServiceClient.getVideo(byId.getVideoId());
+        if (videoResponse == null) {
+            throw new BaseException("Video not found", HttpStatus.NOT_FOUND.value());
+        }
+
         commitRepository.deleteById(id);
     }
 }
